+++ models/contact.js
@@ -1,4 +1,3 @@
-const Joi = require("joi");
 const { Schema, model } = require("mongoose");
 const handleSaveErrors = require("../helpers/handleSaveErrors");
 
@@ -29,20 +28,4 @@ contactSchema.post("save", handleSaveErrors);
 
 const Contact = model("contact", contactSchema);
 
-const addSchema = Joi.object({
-  name: Joi.string().min(2).alphanum().required(),
-  email: Joi.string().trim().email().required(),
-  phone: Joi.number().integer().required(),
-  favorite: Joi.boolean(),
-});
-
-const updateFavoriteSchema = Joi.object({
-  favorite: Joi.boolean().required(),
-});
-
-const schemas = {
-  addSchema,
-  updateFavoriteSchema,
-};
-
-module.exports = { schemas, Contact };
+module.exports = {  Contact };
