+++ schemas/schemas.js
@@ -1,0 +1,19 @@
+const Joi = require("joi");
+
+const addSchema = Joi.object({
+  name: Joi.string().min(2).alphanum().required(),
+  email: Joi.string().trim().email().required(),
+  phone: Joi.number().integer().required(),
+  favorite: Joi.boolean(),
+});
+
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+
+const schemas = {
+  addSchema,
+  updateFavoriteSchema,
+};
+
+module.exports = { schemas };
